+++ COOKBOOK/Views/RecipesDetails/RecipesDetailsView.swift
@@ -95,28 +95,31 @@ class RecipesDetailsView: UIView {
             }
         })
 
-        
-        if button.isSelected == true {
-            let configrations = UIImage.SymbolConfiguration(pointSize: 24)
-            sender.setImage(UIImage(systemName: "heart", withConfiguration: configrations), for: .normal)
-            sender.tintColor = .CustomGreen()
-            sender.backgroundColor = .clear
-            sender.isSelected = false
-            let recipeID = recipeVC.recipeID
-            let uid = Auth.auth().currentUser!.uid
-            db.collection("Users").document(uid).collection("FavouriteRecipes").document(recipeID ?? "Error").delete()
+        if Auth.auth().currentUser?.uid != nil {
+            if button.isSelected == true {
+                let configrations = UIImage.SymbolConfiguration(pointSize: 24)
+                sender.setImage(UIImage(systemName: "heart", withConfiguration: configrations), for: .normal)
+                sender.tintColor = .CustomGreen()
+                sender.backgroundColor = .clear
+                sender.isSelected = false
+                let recipeID = recipeVC.recipeID
+                let uid = Auth.auth().currentUser!.uid
+                db.collection("Users").document(uid).collection("FavouriteRecipes").document(recipeID ?? "Error").delete()
+            } else {
+                let configrations = UIImage.SymbolConfiguration(pointSize: 24)
+                sender.setImage(UIImage(systemName: "heart.fill", withConfiguration: configrations), for: .normal)
+                sender.tintColor = .CustomGreen()
+                sender.backgroundColor = .clear
+                sender.isSelected = true
+                let recipeID = recipeVC.recipeID
+                let uid = Auth.auth().currentUser!.uid
+                let data = [
+                    "FavRecipes": recipeID
+                    ] as! [String: String]
+                db.collection("Users").document(uid).collection("FavouriteRecipes").document(recipeID ?? "Error").setData(data)
+            }
         } else {
-            let configrations = UIImage.SymbolConfiguration(pointSize: 24)
-            sender.setImage(UIImage(systemName: "heart.fill", withConfiguration: configrations), for: .normal)
-            sender.tintColor = .CustomGreen()
-            sender.backgroundColor = .clear
-            sender.isSelected = true
-            let recipeID = recipeVC.recipeID
-            let uid = Auth.auth().currentUser!.uid
-            let data = [
-                "FavRecipes": recipeID
-            ] as! [String: String]
-            db.collection("Users").document(uid).collection("FavouriteRecipes").document(recipeID ?? "Error").setData(data)
+            Alert.showAlert(title: "Error", subtitle: "Please login to save your favorite recipes", leftView: UIImageView(image: #imageLiteral(resourceName: "isErrorIcon")), style: .danger)
         }
     }
